Minor refactoring. Moved GuiTestBase to guitests package since it is only used for integration testing.
