--- src/test/java/address/guitests/PersonEditDialogTest.java
@@ -1,6 +1,0 @@
-package address.guitests;
-
-import address.GuiTestBase;
-
-public class PersonEditDialogTest extends GuiTestBase {
-}
